Extended Access: Create `utils` and `constants` files
